% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geofabrik-package.R
\docType{package}
\name{geofabrik-package}
\alias{geofabrik}
\alias{geofabrik-package}
\title{geofabrik: Downloading Open Street Map Data}
\description{
Download 'OpenStreetMap' data from 'geofabrik' servers <https://download.geofabrik.de/>. This approach uses 
    only the direct link downloads. Besides, this package
    does not import any external package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ibarraespinosa.github.io/geofabrik/}
  \item Report bugs at \url{https://github.com/ibarraespinosa/geofabrik/issues}
}

}
\author{
\strong{Maintainer}: Sergio Ibarra-Espinosa \email{sergio.ibarra@usp.br} (\href{https://orcid.org/0000-0002-3162-1905}{ORCID}) [woodcutter]

}
\keyword{internal}
