% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{urlgeo}
\alias{urlgeo}
\title{Data.frame with links for geofabrik}
\format{
A Spatial data.frame (sf) with 1796 rows and 1 variables:
\describe{
  \item{Region}{Region}
  \item{Subregion}{Subregion}
  \item{x}{region to be used in get_osm}
  \item{LINKPBF}{link}
  \item{size}{size}
  \item{LINKSHP}{link}
  \item{LINKOSM}{link}
  \item{URL_OSMPBF}{URL}
  \item{URL_SHP}{URL}
  \item{URL_OSMBZ2}{URL}
}
}
\source{
\url{https://download.geofabrik.de/}
}
\usage{
urlgeo
}
\description{
Data.frame with links for geofabrik
}
\keyword{datasets}
